+++ bin/build-tools-image.sh
@@ -6,7 +6,7 @@ set -euo pipefail
 #FROM maven:${maven_version}-openjdk-${java_version}-slim
 docker build \
     --build-arg maven_version=3.8.4 \
-    --build-arg java_version=11 \
+    --build-arg java_version=8 \
     -t tools \
     -f tools.Dockerfile \
     .