+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [3.0.6] - 2024-02-20
+
+### Fixed 
+- Restored compatibility with Java 8 (CNJR-7854)
+
 ## [3.0.5] - 2023-06-08
 
 ### Fixed